backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
google-genai>=1.0.0
httpx>=0.27.0
Pillow>=10.3.0
python-multipart>=0.0.9
